rename: _subpixel_utils -> _subpixel_finding
